artifacts/resources/01-ForexPipeline/requirements.txt
-c constraints-3.7.txt
apache-airflow[async,amazon,celery,cncf.kubernetes,docker,dask,elasticsearch,ftp,grpc,hashicorp,http,ldap,google,microsoft.azure,mysql,postgres,redis,sendgrid,sftp,slack,ssh,statsd,virtualenv,apache.hdfs,apache.hive,apache.spark]==2.1.4